Update settings logic and refactor
Update settings logic and refactor
