+++ tests/test_rebar.py
@@ -5,24 +5,24 @@ from mento.material import Concrete_ACI_318_19, SteelBar
 from mento.units import psi, kip, mm, inch, ksi, cm, MPa, kN
 from mento.forces import Forces
 from mento.node import Node
+from mento.settings import BeamSettings
 
 
 @pytest.fixture()
 def beam_example_imperial() -> RectangularBeam:
     concrete = Concrete_ACI_318_19(name="C4", f_c=4000 * psi)
     steelBar = SteelBar(name="ADN 420", f_y=60 * ksi)
-    custom_settings = {
-        "clear_cover": 1.5 * inch,
-        "stirrup_diameter_ini": 0.5 * inch,
-        "longitudinal_diameter_ini": 1 * inch,
-    }
+    beam_settings = BeamSettings(
+        stirrup_diameter_ini=0.5 * inch, minimum_longitudinal_diameter=1 * inch
+    )
     section = RectangularBeam(
         label="V-10x16",
         concrete=concrete,
         steel_bar=steelBar,
         width=10 * inch,
         height=16 * inch,
-        settings=custom_settings,
+        c_c=1.5 * inch,
+        settings=beam_settings,
     )
     return section
 
@@ -31,14 +31,15 @@ def beam_example_imperial() -> RectangularBeam:
 def beam_example_metric() -> RectangularBeam:
     concrete = Concrete_ACI_318_19(name="H30", f_c=30 * MPa)
     steelBar = SteelBar(name="ADN 420", f_y=420 * MPa)
-    custom_settings = {"clear_cover": 30 * mm, "stirrup_diameter_ini": 8 * mm}
+    beam_settings = BeamSettings(stirrup_diameter_ini=8 * mm)
     section = RectangularBeam(
         label="V 20x50",
         concrete=concrete,
         steel_bar=steelBar,
         width=20 * cm,
         height=50 * cm,
-        settings=custom_settings,
+        c_c=30 * mm,
+        settings=beam_settings,
     )
     return section
