+++ tests/test_rectangular.py
@@ -1,15 +1,9 @@
 import pytest
-import matplotlib.pyplot as plt
-from matplotlib.colors import to_rgba
-from matplotlib.patches import Rectangle, FancyBboxPatch
-import numpy as np
 from typing import Generator
 
 from mento.material import SteelBar, Concrete
 from mento.units import inch, cm, mm, MPa
 from mento.rectangular import RectangularSection  # Assuming this is the module name
-from mento.settings import GLOBAL_BEAM_SETTINGS
-from mento.results import CUSTOM_COLORS
 
 
 @pytest.fixture()
@@ -82,135 +76,3 @@ def test_rectangular_section_geometric_properties(
 
     expected_I_z = (50 * cm) * (30 * cm) ** 3 / 12
     assert basic_rectangular_section.I_z == expected_I_z.to("cm**4")
-
-
-def test_rectangular_section_settings_property(
-    basic_rectangular_section: RectangularSection,
-) -> None:
-    """Test that the settings property returns GLOBAL_BEAM_SETTINGS."""
-    assert basic_rectangular_section.settings is GLOBAL_BEAM_SETTINGS
-
-
-# --- Test for the plot method ---
-
-
-def test_rectangular_section_plot_smoke_test(
-    basic_rectangular_section: RectangularSection,
-) -> None:
-    """
-    Smoke test for the plot method.
-    Simply calls the method and ensures no exceptions are raised.
-    Closes the plot to prevent it from showing during tests.
-    """
-    try:
-        basic_rectangular_section.plot()
-        # Close the plot to prevent it from popping up
-        plt.close()
-    except Exception as e:
-        pytest.fail(f"Plot method raised an exception: {e}")
-
-
-def test_rectangular_section_plot_components(
-    basic_rectangular_section: RectangularSection,
-) -> None:
-    """
-    Test that the plot method adds the expected matplotlib components
-    (main rectangle, stirrup rectangles, annotations).
-    """
-    basic_rectangular_section.plot()
-    ax = basic_rectangular_section._ax  # Access the private _ax created by plot
-
-    assert ax is not None, "Plot method should create and assign an Axes object."
-
-    # Check for patches (rectangles, FancyBboxPatch)
-    # Expecting: 1 main Rectangle, 2 FancyBboxPatches for stirrup
-    # + potentially other internal patches depending on matplotlib's rendering
-    # It's safer to check for specific types and properties if possible.
-
-    # Filter for Rectangle and FancyBboxPatch
-    rectangles = [p for p in ax.patches if isinstance(p, Rectangle)]
-    fancy_bboxes = [p for p in ax.patches if isinstance(p, FancyBboxPatch)]
-
-    assert (
-        len(rectangles) >= 1
-    ), "Expected at least one Rectangle patch (the main section)."
-    assert (
-        len(fancy_bboxes) == 2
-    ), "Expected exactly two FancyBboxPatch objects for the stirrup."
-
-    # Check for annotations (dimensions)
-    # Annotations are stored in ax.texts for text, ax.lines or ax.artists for arrows
-    # matplotlib.axes.Axes.annotate returns an Annotation object.
-    # Text labels are found in ax.texts
-    assert (
-        len(ax.texts) == 4
-    ), "Expected 4 text annotations for dimensions (width, height)."
-
-    # You could add more specific checks, e.g.:
-    # - Check the coordinates of the main rectangle:
-    main_rect = next((p for p in rectangles if p.get_x() == 0 and p.get_y() == 0), None)
-    assert main_rect is not None
-    assert np.isclose(main_rect.get_width(), basic_rectangular_section.width.magnitude)
-    assert np.isclose(
-        main_rect.get_height(), basic_rectangular_section.height.magnitude
-    )
-    assert main_rect.get_edgecolor() == to_rgba(CUSTOM_COLORS["dark_gray"])
-
-    # - Check stirrup patch properties (example, will need exact dimensions based on calculation)
-    #   This gets more complex quickly, focusing on existence and basic properties is often enough.
-
-    plt.close()  # Close the figure created by the plot method
-
-
-def test_rectangular_section_plot_imperial_units_text(
-    imperial_rectangular_section: RectangularSection,
-) -> None:
-    """
-    Test that the plot method displays dimensions in imperial units
-    when the concrete's unit system is set to 'imperial'.
-    """
-    imperial_rectangular_section.plot()
-    ax = imperial_rectangular_section._ax
-    assert ax is not None, "The plot method did not assign an Axes object to _ax."
-
-    # Extract text labels
-    text_labels = [t.get_text() for t in ax.texts]
-
-    # Check if the labels contain "inch" (represented by "in" in pint's compact format)
-    # Adjust based on how pint formats "inch" in your setup (e.g., "in", "inch", etc.)
-    # "{:~P}" often uses common abbreviations.
-    assert any(
-        "in" in label for label in text_labels
-    ), "Expected imperial unit 'in' in dimension text."
-    assert (
-        "{:.0f~P}".format(imperial_rectangular_section.width.to("inch")) in text_labels
-    )
-    assert (
-        "{:.0f~P}".format(imperial_rectangular_section.height.to("inch")) in text_labels
-    )
-
-    plt.close()
-
-
-def test_rectangular_section_plot_metric_units_text(
-    basic_rectangular_section: RectangularSection,
-) -> None:
-    """
-    Test that the plot method displays dimensions in metric units (cm)
-    when the concrete's unit system is not 'imperial'.
-    """
-    basic_rectangular_section.plot()  # Ensure the plot method is called
-    ax = basic_rectangular_section._ax
-    assert ax is not None, "The plot method did not assign an Axes object to _ax."
-
-    # Extract text labels
-    text_labels = [t.get_text() for t in ax.texts]
-
-    # Check if the labels contain "cm"
-    assert any(
-        "cm" in label for label in text_labels
-    ), "Expected metric unit 'cm' in dimension text."
-    assert "{:.0f~P}".format(basic_rectangular_section.width.to("cm")) in text_labels
-    assert "{:.0f~P}".format(basic_rectangular_section.height.to("cm")) in text_labels
-
-    plt.close()
